--- client/src/components/UserAccountComponent/ChangeCardModal/index.js
@@ -1,1 +1,0 @@
-export { default } from './ChangeCardModal';
