+++ client/src/components/UserAccountComponent/CommentsUserSend/index.js
@@ -1,0 +1,1 @@
+export { default } from './CommentsUserSend';
