+++ client/src/components/UserAccountComponent/MyTrips/ChangeTrip/index.js
@@ -1,0 +1,1 @@
+export { default } from './ChangeTrip';
