+++ client/src/components/UserAccountComponent/TravelWith/index.js
@@ -1,0 +1,1 @@
+export { default } from './TravelWith';
