+++ client/src/components/UserAccountComponent/UserAccount/UserAccount.jsx
@@ -7,7 +7,8 @@ import { ListGroup, ListGroupItem } from 'reactstrap';
 import { useSelector, useDispatch } from 'react-redux';
 import { useParams } from 'react-router-dom';
 import { allInfo } from '../../../redux/actions/userActions';
-import UserAccountTrips from '../UserAccountTrips';
+import UserAccountTrips from '../MyTrips/UserAccountTrips';
+import CommentsUserSend from '../CommentsUserSend';
 
 export default function UserAccount() {
   const dispatch = useDispatch();
@@ -101,6 +102,7 @@ export default function UserAccount() {
         </Grid>
       </Grid>
       <UserAccountTrips />
+      <CommentsUserSend />
     </Container>
   );
 }
